refactor(brett): extract admin commands to ws-admin-commands.ts (S1 gate fix) [T000459]
ws-handler.ts exceeded the S1 quality gate limit (645 lines > 600).
Moves handleAssignRole + admin switch into ws-admin-commands.ts,
reducing ws-handler.ts to 472 lines. Re-export preserves the public API.
